+++ src/mock/event-mock.js
@@ -1,6 +1,8 @@
 import {CityList, OfferList, TransferList, ActivityList} from '../const.js';
-import {getRandomInteger, getRandomArrayItem, getRandomDateDay, cloneArray} from '../utils.js';
-
+import {getRandomInteger} from '../utils/common.js';
+import {getRandomArrayItem} from '../utils/common.js';
+import {cloneArray} from '../utils/clone-method.js';
+import {getRandomDateDay} from '../utils/date.js';
 
 const generateDescription = () => {
   const DescriptionText = `Lorem ipsum dolor sit amet, consectetur adipiscing elit. Cras aliquet varius magna, non porta ligula feugiat eget. Fusce tristique felis at fermentum pharetra. Aliquam id orci ut lectus varius viverra. Nullam nunc ex, convallis sed finibus eget, sollicitudin eget ante. Phasellus eros mauris, condimentum sed nibh vitae, sodales efficitur ipsum. Sed blandit, eros vel aliquam faucibus, purus ex euismod diam, eu luctus nunc ante ut dui. Sed sed nisi sed augue convallis suscipit in sed felis. Aliquam erat volutpat. Nunc fermentum tortor ac porta dapibus. In rutrum ac purus sit amet tempus.`;
