+++ src/utils/date.js
@@ -1,12 +1,4 @@
-export const getRandomInteger = (min, max) => {
-  return min + Math.floor(Math.random() * (max - min + 1));
-};
-
-
-export const getRandomArrayItem = (array, startIndex = 0) => {
-  return array[getRandomInteger(startIndex, array.length - 1)];
-};
-
+import {getRandomInteger} from './common.js';
 
 const getRandomHours = (date, intervalHours) => {
   const targetDate = new Date(date);
@@ -15,6 +7,7 @@ const getRandomHours = (date, intervalHours) => {
   targetDate.setHours(targetDate.getHours() + hours, targetDate.getMinutes() + minutes);
   return targetDate;
 };
+
 
 export const getRandomDateDay = (referencePoint = null, startDayEvent = 3, intervalHours = 8) => {
   let dateStart = 0;
@@ -34,6 +27,7 @@ export const getRandomDateDay = (referencePoint = null, startDayEvent = 3, inter
   return {dateStart, dateEnd};
 };
 
+
 export const castTimeFormat = (value) => {
   return value < 10 ? `0${value}` : value;
 };
@@ -46,6 +40,7 @@ export const formatTime = (date) => {
   return `${hours}:${minutes}`;
 };
 
+
 export const formatDateTime = (date) => {
   const yyyy = date.getFullYear();
   const mm = castTimeFormat(date.getMonth() + 1);
@@ -57,68 +52,4 @@ export const formatDateTime = (date) => {
     date: `${yyyy}-${mm}-${dd}`,
     dateTime: `${dd}/${mm}/${String(yyyy).slice(2)} ${hours}:${minutes}`,
   };
-};
-
-const isObject = (object) => {
-  const type = typeof object;
-  return type === `function` || type === `object`;
-};
-
-const cloneObject = (source) => {
-  const clone = {};
-
-  for (const key in source) {
-    if (source.hasOwnProperty(key)) {
-      if (isObject(source[key])) {
-        clone[key] = cloneObject(source[key]);
-      } else {
-        clone[key] = source[key];
-      }
-    }
-  }
-  return clone;
-};
-
-export const cloneArray = (source) => {
-  const clone = [];
-
-  for (const key in source) {
-    if (source.hasOwnProperty(key)) {
-      if (isObject(source[key])) {
-        clone[key] = cloneObject(source[key]);
-      } else {
-        clone[key] = source[key];
-      }
-    }
-  }
-  return clone;
-};
-
-
-export const createElement = (template) => {
-  const newElement = document.createElement(`div`);
-  newElement.innerHTML = template;
-
-  return newElement.firstElementChild;
-};
-
-
-export const RenderPosition = {
-  AFTERBEGIN: `afterbegin`,
-  BEFOREEND: `beforeend`,
-  AFTEREND: `afterend`
-};
-
-export const render = (container, element, place) => {
-  switch (place) {
-    case RenderPosition.AFTERBEGIN:
-      container.prepend(element);
-      break;
-    case RenderPosition.BEFOREEND:
-      container.append(element);
-      break;
-    case RenderPosition.AFTEREND:
-      container.after(element);
-      break;
-  }
 };
